Viewer: Separate layout from main window
