+++ apps/desktop/src/routes/(main)/+page.svelte
