+++ apps/desktop/src/routes/viewer/+layout.svelte
@@ -1,0 +1,2 @@
+<!-- Viewer uses root layout (CSS + logger). No updater, notifications, or window state. -->
+<slot />
